Add dependencies for integration tests to build-deps.sh & enable in CI.
These dependencies are required to run the integration tests on the
Ubuntu instance used by travis-ci (xenial). A more current distro (like
Debian testing / buster) will be able to install them from packages.

This commit removes the use of the gnu-efi packaged in the travis-ci
build image building it the 3.0.9 source release instead. We require a
version > 3.0.8.

Clang won't build the master branch of swtpm currently. This commit
includes a workaround that forces the use of gcc for this component
only.

+++ .ci/build-deps.sh
@@ -65,5 +65,68 @@ $MAKE
 sudo $MAKE install
 cd -
 
+git clone --branch master --single-branch \
+  https://github.com/stefanberger/libtpms
+cd libtpms
+./autogen.sh --with-openssl --with-tpm2 --prefix=${PREFIX}
+$MAKE
+sudo $MAKE install
+cd -
+
+# swtpm isn't compatible with clang
+CC_BAK=$CC
+CC=gcc
+git clone --branch master --single-branch \
+  https://github.com/stefanberger/swtpm
+cd swtpm
+./autogen.sh --prefix=${PREFIX}
+$MAKE
+sudo $MAKE install
+cd -
+CC=$CC_BAK
+
+QEMU_SRC=qemu-3.1.0.tar.xz
+QEMU_SRC_SHA256=6a0508df079a0a33c2487ca936a56c12122f105b8a96a44374704bef6c69abfc
+wget "https://download.qemu.org/${QEMU_SRC}"
+sha256sum "${QEMU_SRC}" | grep -q "${QEMU_SRC_SHA256}"
+if [ ! $? -eq 0 ]; then
+    echo "hash mismatch on qemu tarball"
+    exit 1
+fi
+tar axf ${QEMU_SRC}
+cd qemu-3.1.0
+./configure --prefix=${PREFIX} --enable-tpm --disable-kvm \
+  --target-list=x86_64-softmmu
+$MAKE
+sudo $MAKE install
+cd -
+
+GNUEFI_BZ2=gnu-efi-3.0.9.tar.bz2
+GNUEFI_BZ2_SHA256=6715ea7eae1c7e4fc5041034bd3f107ec2911962ed284a081e491646b12277f0
+wget "https://downloads.sourceforge.net/project/gnu-efi/${GNUEFI_BZ2}"
+sha256sum "${GNUEFI_BZ2}" | grep -q "${GNUEFI_BZ2_SHA256}"
+if [ ! $? -eq 0 ]; then
+    echo "hash mismatch on gnu-efi tarball"
+    exit 1
+fi
+tar axf ${GNUEFI_BZ2}
+cd gnu-efi-3.0.9
+make
+sudo $MAKE PREFIX=${PREFIX} install
+cd -
+
+OVMF_RPM=edk2.git-ovmf-x64-0-20190131.914.g7a90895306.noarch.rpm
+OVMF_RPM_SHA256=582975520720e318c210c6930d046929d35c703e4ec275e37c8e1e89737d345d
+wget "https://www.kraxel.org/repos/jenkins/edk2/${OVMF_RPM}"
+sha256sum "${OVMF_RPM}" | grep -q "${OVMF_RPM_SHA256}"
+if [ ! $? -eq 0 ]; then
+    echo "hash mismatch on OVMF rpm"
+    exit 1
+fi
+rpm2cpio "${OVMF_RPM}" | cpio -i -d --no-absolute-filenames -v
+OVMF_FD=$(find ./ -name 'OVMF-pure-efi.fd')
+sudo mkdir /usr/share/ovmf
+sudo cp ${OVMF_FD} /usr/share/ovmf/OVMF.fd
+
 cd ${STARTDIR}
 rm -rf "${WORKDIR}"
